remove all use of sscanf, fix PORT poll bug, other misc changes.
removed all uses of sscanf as it's quite slow and the size spec was clunky to use.
in most cases, sscanf was replaced with snprintf and strtol in others.
the reason for sscanf was the input buffer not being null terminated, instead TELNET_EOL terminated.
however, i can just simply replace the TELNET_EOL with null...which is what i did.

polling for port was broken due to using the wrong poll type POLLOUT rather than POLLIN.

pollfds array was made static as it can be rather large with many ftp sessions.

better error messages in RETR and STOR by returning the fullpath as part of the error message.
this will help track down any issues with clients.

in a few cases, the result of snprintf only checked if the value was greater than the output buffer.
this was incorrect. if the result is >= buffer size, then that indicates the buffer was too small for the output.