add session timeout, refactor, remove sendfile
session timeout was added for clients that don't properly clean state, ie, leave sessions open.
as ftpsrv was designed for a fixed number of sessions, this can be an issue if the client
consumes all sessions. this would require re-starting ftpsrv.
now, all clients i have used behave properly, but just in case, ive added this.

remove sendfile, for now at least. its only used on linux, and whilst it's useful for me as i can acheive 500MiB/s
upload with it, no other platform has it. on pc, without sendfile, i get ~200MiB/s which is still plenty
fast enough.

big refactor, mainly just for code-style reasons, it's functionally the same as before.

nds sockets need to be closed with closesocket, not close. this is now fixed.