add timeout for data transfer loop, Nx mounting bis, hdd, save_system, save_cache.
the data transfer loop now has a timeout of 1ms. this was added based on the theory that
if the disk speed was slower than the network, then the data loop would block until
the entire file / directory is sent.
usually this wouldnt be the case, until i was transfering files from a hdd.
the seek time can be pretty bad, and thus the network is faster.

many new nx mount options were added, the biggest one being hdd. the impl isn't perfect yet, and it
can block the main thread for quite a while as the disk starts to spin up for the first time.
however i decided that blocking is the best as otherwise opening a folder would return an
error as it isn;t ready yet, however to the user, there is no clear reason as to why it didnt open.
by blocking, the user will likely hear the disk spinning up and flashing, and understand why its
taking some time.